chore: move utils outside src
